squash
fix unit tests

cleanups

renames

unit tests

interface modifications

linter

linter

linter

+++ rpc/v9/subscriptions_test.go
@@ -91,10 +91,14 @@ func (fs *fakeSyncer) HighestBlockHeader() *core.Header {
 func (fs *fakeSyncer) PendingData() (core.PendingData, error) {
 	return nil, core.ErrPendingDataNotFound
 }
-func (fs *fakeSyncer) PendingBlock() *core.Block                             { return nil }
-func (fs *fakeSyncer) PendingState() (core.StateReader, func() error, error) { return nil, nil, nil }
+func (fs *fakeSyncer) PendingBlock() *core.Block { return nil }
+func (fs *fakeSyncer) PendingState() (core.StateReader, func() error, error) {
+	return nil, nil, nil
+}
 
-func (fs *fakeSyncer) PendingStateBeforeIndex(index int) (core.StateReader, func() error, error) {
+func (fs *fakeSyncer) PendingStateBeforeIndex(
+	index uint64,
+) (core.StateReader, func() error, error) {
 	return nil, nil, nil
 }
